requirements.txt
pyautogui==0.9.53
easyocr==1.5.0
pillow==9.2.0
opencv-python==4.6.0.66
numpy==1.23.1
requests==2.28.1